pyproject.toml
[tool.poetry]
name = "elevatus-assignment-py"
version = "0.1.0"
description = "Technical Assignment by Elevatus for a Python Developer Role"
authors = ["Moha-AlHanbali <Moha.AlHanbali@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
fastapi = "^0.108.0"
pymongo = "^4.6.1"
uvicorn = "^0.25.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
setuptools = "^69.0.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
httpx = "^0.26.0"
pytest-mongodb = "^2.4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"